Updated example

+++ Buildings/Resources/Python-Sources/testFunctions.py
@@ -36,7 +36,7 @@ def s2_r1(xS):
     return y
 
 def r1i1_r2i1(xR, xI):
-    return [[2.*xR, 2.*float(xI)], 3]
+    return [[2.*xR, 2.*float(xI)], long(3)]
 
 # Functions with memory
 def r1_r1PassPythonObject(xR, obj):
